Move util to correct spot, as per #1630
